Simplify local file handling by integrating with IncludeFile for now; revert to deploy_mode
Also contains a few other fixes to address comments
