+++ metaflow/datatools/__init__.py
@@ -1,6 +1,2 @@
 
 from .s3 import S3
-from .local import Local
-
-DATACLIENTS = {'local': Local,
-               's3': S3}
